chat/requirements.txt
fastapi
fastapi[standard]
uvicorn
uvicorn[standard]
typing
python-jose